+++ jdi-bdd-tests/src/main/java/io/github/com/pages/GithubPage.java
@@ -4,7 +4,7 @@ import com.epam.jdi.light.ui.html.elements.common.Text;
 import org.openqa.selenium.support.FindBy;
 
 public class GithubPage {
-    @FindBy(css = "[itemprop=about]")
+    @FindBy(css = ".BorderGrid-cell p.f4")
     public static Text repoDescription;
     // @UI(".repository-meta-content")
     // public static UIElement repoDescription;