+++ jdi-light-html-tests/src/main/java/io/github/com/pages/SimpleTablePage.java
@@ -1,14 +1,14 @@
 package io.github.com.pages;
 
 import com.epam.jdi.light.elements.complex.table.DataTable;
-import com.epam.jdi.light.elements.complex.table.Table;
+import com.epam.jdi.light.elements.complex.table.Grid;
 import com.epam.jdi.light.elements.composite.WebPage;
 import com.epam.jdi.light.elements.pageobjects.annotations.locators.JTable;
 import io.github.com.entities.Furniture;
 
 public class SimpleTablePage extends WebPage {
 
-    public static Table simpleTable;
+    public static Grid simpleTable;
     public static DataTable<?, Furniture> products;
 
     @JTable(