+++ jdi-light/src/main/java/com/epam/jdi/light/elements/complex/DataList.java
@@ -4,7 +4,6 @@ import com.epam.jdi.light.asserts.core.DataListAssert;
 import com.epam.jdi.light.common.JDIAction;
 import com.epam.jdi.light.elements.interfaces.base.ICoreElement;
 import com.epam.jdi.tools.LinqUtils;
-import com.epam.jdi.tools.PrintUtils;
 import org.hamcrest.Matcher;
 import org.hamcrest.MatcherAssert;
 
@@ -16,6 +15,7 @@ import static com.epam.jdi.light.asserts.core.SoftAssert.assertSoft;
 import static com.epam.jdi.light.common.Exceptions.exception;
 import static com.epam.jdi.light.common.UIUtils.asEntity;
 import static com.epam.jdi.tools.EnumUtils.getEnumValue;
+import static com.epam.jdi.tools.PrintUtils.print;
 import static com.epam.jdi.tools.ReflectionUtils.getGenericTypes;
 
 /**
@@ -48,7 +48,7 @@ public class DataList<T extends ICoreElement, D> extends ListBase<T, DataListAss
     }
     @Override
     public String getValue() {
-        return PrintUtils.print(LinqUtils.map(asData(), Object::toString));
+        return print(asData(), Object::toString);
     }
     
     @Override
@@ -88,5 +88,9 @@ public class DataList<T extends ICoreElement, D> extends ListBase<T, DataListAss
         } catch (Exception ignore) {
 
         }
+    }
+    @Override
+    public List<String> values() {
+        return LinqUtils.map(asData(), Object::toString);
     }
 }