+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ## 1.2.16 (release 17.05.2020)
 * Add keyboard commands (press, command, commands, pasteText) to UIElement
+* Add option to take screenshot via Robot
 
 ## 1.2.15 (release 08.06.2020)
 * Fix list issue with multiple same values